requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.109.1
pinecone==5.4.0
requests==2.31.0
PyMuPDF==1.24.10
Pillow==10.4.0
pytesseract==0.3.10
tiktoken==0.7.0